Create dz 3

+++ OOP/Blackjack/blackjack.cpp
@@ -55,7 +55,7 @@ void print_cards(std::vector<PlayingCard> cards)
         std::cout << "No cards!" << std::endl;
         return;
     }
-    for (int i = 0; i < cards.size() - 1; i++)
+    for (size_t i = 0; i < cards.size() - 1; i++)
     {
         PlayingCard card = cards[i];
         if (card.is_hidden())
